adding cobra and some refactoring
