+++ cmd/clusters-service/main.go
@@ -17,43 +17,40 @@ limitations under the License.
 package main
 
 import (
-	"fmt"
+	"flag"
 	"os"
 
-	"github.com/container-mgmt/dedicated-portal/pkg/signals"
-	"github.com/container-mgmt/dedicated-portal/pkg/sql"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 )
 
-func main() {
-	// Set up signals so we handle the first shutdown signal gracefully:
-	stopCh := signals.SetupHandler()
-	url := ConnectionURL()
-	err := sql.EnsureSchema(
-		"/usr/local/share/clusters-service/migrations",
-		url,
-	)
-	if err != nil {
-		panic(err)
+var (
+	// Main command:
+	rootCmd = &cobra.Command{
+		Use:  "clusters-service",
+		Long: "A service to provision and manage openshift clusteres.",
 	}
-	service := NewClustersService(url)
-	fmt.Println("Created cluster service.")
+)
 
-	// This is temporary and should be replaced with reading from the queue
-	server := NewServer(stopCh, service)
-	err = server.start()
-	if err != nil {
-		panic(fmt.Sprintf("Error starting server: %v", err))
-	}
-	fmt.Println("Created server.")
+func init() {
+	// Send logs to the standard error stream by default:
+	flag.Set("logtostderr", "true")
 
-	fmt.Println("Waiting for stop signal")
-	<-stopCh // wait until requested to stop.
+	// Register the options that are managed by the 'flag' package, so that they will also be parsed
+	// by the 'pflag' package:
+	pflag.CommandLine.AddGoFlagSet(flag.CommandLine)
+
+	// Register the subcommands:
+	rootCmd.AddCommand(serveCmd)
 }
 
-// ConnectionURL generates a connection string from the environment.
-func ConnectionURL() string {
-	return fmt.Sprintf("postgres://%s:%s@localhost:5432/%s?sslmode=disable",
-		os.Getenv("POSTGRESQL_USER"),
-		os.Getenv("POSTGRESQL_PASSWORD"),
-		os.Getenv("POSTGRESQL_DATABASE"))
+func main() {
+	// This is needed to make `glog` believe that the flags have already been parsed, otherwise
+	// every log messages is prefixed by an error message stating the the flags haven't been
+	// parsed.
+	flag.CommandLine.Parse([]string{})
+
+	// Execute the root command:
+	rootCmd.SetArgs(os.Args[1:])
+	rootCmd.Execute()
 }
